Improve float handling Fixed YaLTeR/niri/pull/2108

+++ src/niri.rs
@@ -3177,14 +3177,10 @@ impl Niri {
             let size = transform.transform_size(output_size);
 
             let hot_top_left = Rectangle::new(Point::new(0., 0.), Size::from((1., 1.)));
-            let hot_top_right =
-                Rectangle::new(Point::new((size.w - 1) as f64, 0.), Size::from((1., 1.)));
-            let hot_bottom_left =
-                Rectangle::new(Point::new(0., (size.h - 1) as f64), Size::from((1., 1.)));
-            let hot_bottom_right = Rectangle::new(
-                Point::new((size.w - 1) as f64, (size.h - 1) as f64),
-                Size::from((1., 1.)),
-            );
+            let hot_top_right = Rectangle::new(Point::new(size.w - 1., 0.), Size::from((1., 1.)));
+            let hot_bottom_left = Rectangle::new(Point::new(0., size.h - 1.), Size::from((1., 1.)));
+            let hot_bottom_right =
+                Rectangle::new(Point::new(size.w - 1., size.h - 1.), Size::from((1., 1.)));
 
             if !(hot_corners.top_left
                 || hot_corners.top_right
